requirements.txt
streamlit
pandas
numpy>
pdfplumber
PyPDF2
requests
beautifulsoup4
python-dotenv
google-generativeai
Pillow